requirements_dev.txt
-r requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
astroid==2.14.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
black==23.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dill==0.3.6
django-cors-headers==3.13.0
django_debug_toolbar==3.8.1
docopt==0.6.2
frozenlist==1.3.3
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.1.1
packaging==23.0
pathspec==0.11.0
pep8==1.7.1
pipreqs==0.4.11
platformdirs==2.6.2
pycodestyle==2.10.0
pylint==2.16.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.14
wrapt==1.14.1
yarg==0.1.9
yarl==1.8.2
